Template Scene
A template scene to build the world scenes. Should be set up with everything. If something is missing feel free to update the template, but know it wont update the rest of the stuff. This should make it alot easier to set up rooms (Just delete the player possibly. Not 100% sure what all is actually persistant between scenes to be conscious of that. More likely than not the audio will need to be attached to the player for world audio in order to not have it restart every scene.)